// npm-packages/tests/udf-tests/convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  action: typeof action;
  adversarial: typeof adversarial;
  adversarialWasm: typeof adversarialWasm;
  analyze: typeof analyze;
  args_validation: typeof args_validation;
  asyncTests: typeof asyncTests;
  auth: typeof auth;
  basic: typeof basic;
  benches: typeof benches;
  creationTime: typeof creationTime;
  crons: typeof crons;
  crons_error: typeof crons_error;
  custom_errors: typeof custom_errors;
  "directory/defaultTest": typeof directory_defaultTest;
  "directory/udfs": typeof directory_udfs;
  environmentVariables: typeof environmentVariables;
  fetch: typeof fetch;
  globals: typeof globals;
  helpers: typeof helpers;
  http: typeof http;
  http_action: typeof http_action;
  http_all_imported_handlers: typeof http_all_imported_handlers;
  http_no_default: typeof http_no_default;
  http_no_imports: typeof http_no_imports;
  http_object_default: typeof http_object_default;
  http_undefined_default: typeof http_undefined_default;
  idEncoding: typeof idEncoding;
  idStrings: typeof idStrings;
  import_tests: typeof import_tests;
  indexing: typeof indexing;
  internal: typeof internal_;
  js: typeof js;
  "js_builtins/abort_controller": typeof js_builtins_abort_controller;
  "js_builtins/blob": typeof js_builtins_blob;
  "js_builtins/crypto": typeof js_builtins_crypto;
  "js_builtins/event": typeof js_builtins_event;
  "js_builtins/event_target": typeof js_builtins_event_target;
  "js_builtins/file": typeof js_builtins_file;
  "js_builtins/headers": typeof js_builtins_headers;
  "js_builtins/request": typeof js_builtins_request;
  "js_builtins/response": typeof js_builtins_response;
  "js_builtins/setTimeout": typeof js_builtins_setTimeout;
  "js_builtins/stream": typeof js_builtins_stream;
  "js_builtins/structuredClone": typeof js_builtins_structuredClone;
  "js_builtins/testHelpers": typeof js_builtins_testHelpers;
  "js_builtins/textEncoder": typeof js_builtins_textEncoder;
  "js_builtins/url": typeof js_builtins_url;
  "js_builtins/urlSearchParams": typeof js_builtins_urlSearchParams;
  load_failure: typeof load_failure;
  logging: typeof logging;
  name: typeof name;
  node_actions: typeof node_actions;
  query: typeof query;
  returns_validation: typeof returns_validation;
  scheduler: typeof scheduler;
  search: typeof search;
  shapes: typeof shapes;
  size_errors: typeof size_errors;
  sourceMaps: typeof sourceMaps;
  storage: typeof storage;
  sync: typeof sync;
  transactionMetrics: typeof transactionMetrics;
  unicode: typeof unicode;
  userError: typeof userError;
  values: typeof values;
  vector_search: typeof vector_search;
  wasmTests: typeof wasmTests;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;

/**
 * A utility for referencing Convex functions in your app's internal API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = internal.myModule.myFunction;
 * ```
 */
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;